[Unit] Restore promotion creating tests
