+++ disabledTests/Api/Admin/PromotionsTest.php
@@ -73,378 +73,6 @@ final class PromotionsTest extends JsonApiTestCase
     }
 
     /** @test */
-    public function it_creates_promotion(): void
-    {
-        $this->loadFixturesFromFiles([
-            'authentication/api_administrator.yaml',
-            'channel.yaml',
-            'country.yaml',
-            'customer_group.yaml',
-            'promotion/product.yaml',
-            'promotion/taxon.yaml',
-        ]);
-        $header = array_merge($this->logInAdminUser('api@example.com'), self::CONTENT_TYPE_HEADER);
-
-        $this->client->request(
-            method: 'POST',
-            uri: '/api/v2/admin/promotions',
-            server: $header,
-            content: json_encode([
-                'name' => 'T-Shirts discount',
-                'code' => 'tshirts_discount',
-                'channels' => [
-                    '/api/v2/admin/channels/WEB',
-                ],
-                'description' => 'T-Shirts discount',
-                'priority' => 22,
-                'appliesToDiscounted' => false,
-                'exclusive' => true,
-                'usageLimit' => 3,
-                'couponBased' => true,
-                'startsAt' => '2023-10-04 12:30:00',
-                'endsAt' => '2023-11-04 12:30:00',
-                'translations' => ['en_US' => [
-                    'label' => 'T-Shirts discount',
-                ]],
-                'rules' => [
-                    [
-                        'type' => CartQuantityRuleChecker::TYPE,
-                        'configuration' => [
-                            'count' => 6,
-                        ],
-                    ],
-                    [
-                        'type' => CustomerGroupRuleChecker::TYPE,
-                        'configuration' => [
-                            'group_code' => 'vip',
-                        ],
-                    ],
-                    [
-                        'type' => NthOrderRuleChecker::TYPE,
-                        'configuration' => [
-                            'nth' => 2,
-                        ],
-                    ],
-                    [
-                        'type' => ShippingCountryRuleChecker::TYPE,
-                        'configuration' => [
-                            'country' => 'US',
-                        ],
-                    ],
-                    [
-                        'type' => HasTaxonRuleChecker::TYPE,
-                        'configuration' => [
-                            'taxons' => ['MUGS', 'CAPS'],
-                        ],
-                    ],
-                    [
-                        'type' => TotalOfItemsFromTaxonRuleChecker::TYPE,
-                        'configuration' => [
-                            'WEB' => [
-                                'taxon' => 'MUGS',
-                                'amount' => 1000,
-                            ],
-                            'MOBILE' => [
-                                'taxon' => 'CAPS',
-                                'amount' => 2000,
-                            ],
-                        ],
-                    ],
-                    [
-                        'type' => ContainsProductRuleChecker::TYPE,
-                        'configuration' => [
-                            'product_code' => 'CAP',
-                        ],
-                    ],
-                    [
-                        'type' => ItemTotalRuleChecker::TYPE,
-                        'configuration' => [
-                            'WEB' => [
-                                'amount' => 333,
-                            ],
-                            'MOBILE' => [
-                                'amount' => 222,
-                            ],
-                        ],
-                    ],
-                ],
-                'actions' => [
-                    [
-                        'type' => FixedDiscountPromotionActionCommand::TYPE,
-                        'configuration' => [
-                            'WEB' => [
-                                'amount' => 1000,
-                            ],
-                            'MOBILE' => [
-                                'amount' => 2000,
-                            ],
-                        ],
-                    ],
-                    [
-                        'type' => UnitFixedDiscountPromotionActionCommand::TYPE,
-                        'configuration' => [
-                            'WEB' => [
-                                'amount' => 3000,
-                                'filters' => [
-                                    'price_range_filter' => [
-                                        'min' => 1000,
-                                        'max' => 10000,
-                                    ],
-                                ],
-                            ],
-                            'MOBILE' => [
-                                'amount' => 4000,
-                            ],
-                        ],
-                    ],
-                    [
-                        'type' => PercentageDiscountPromotionActionCommand::TYPE,
-                        'configuration' => [
-                            'percentage' => 0.5,
-                        ],
-                    ],
-                    [
-                        'type' => UnitPercentageDiscountPromotionActionCommand::TYPE,
-                        'configuration' => [
-                            'WEB' => [
-                                'percentage' => 0.1,
-                            ],
-                            'MOBILE' => [
-                                'percentage' => 1,
-                            ],
-                        ],
-                    ],
-                    [
-                        'type' => ShippingPercentageDiscountPromotionActionCommand::TYPE,
-                        'configuration' => [
-                            'percentage' => 0.2,
-                        ],
-                    ],
-                ],
-            ], \JSON_THROW_ON_ERROR),
-        );
-
-        $this->assertResponse(
-            $this->client->getResponse(),
-            'admin/promotion/post_promotion_response',
-            Response::HTTP_CREATED,
-        );
-    }
-
-    /** @test */
-    public function it_does_not_create_a_promotion_without_required_data(): void
-    {
-        $this->loadFixturesFromFiles(['authentication/api_administrator.yaml']);
-        $header = array_merge($this->logInAdminUser('api@example.com'), self::CONTENT_TYPE_HEADER);
-
-        $this->client->request(
-            method: 'POST',
-            uri: '/api/v2/admin/promotions',
-            server: $header,
-            content: json_encode([], \JSON_THROW_ON_ERROR),
-        );
-
-        $this->assertResponse(
-            $this->client->getResponse(),
-            'admin/promotion/post_promotion_without_required_data_response',
-            Response::HTTP_UNPROCESSABLE_ENTITY,
-        );
-    }
-
-    /** @test */
-    public function it_does_not_create_a_promotion_with_taken_code(): void
-    {
-        $this->loadFixturesFromFiles(['authentication/api_administrator.yaml', 'channel.yaml', 'promotion/promotion.yaml']);
-        $header = array_merge($this->logInAdminUser('api@example.com'), self::CONTENT_TYPE_HEADER);
-
-        $this->client->request(
-            method: 'POST',
-            uri: '/api/v2/admin/promotions',
-            server: $header,
-            content: json_encode([
-                'name' => '50% Off on your first order',
-                'code' => '50_off',
-            ], \JSON_THROW_ON_ERROR),
-        );
-
-        $this->assertResponse(
-            $this->client->getResponse(),
-            'admin/promotion/post_promotion_with_taken_code_response',
-            Response::HTTP_UNPROCESSABLE_ENTITY,
-        );
-    }
-
-    /** @test */
-    public function it_does_not_create_a_promotion_with_end_date_earlier_than_start_date(): void
-    {
-        $this->loadFixturesFromFiles(['authentication/api_administrator.yaml']);
-        $header = array_merge($this->logInAdminUser('api@example.com'), self::CONTENT_TYPE_HEADER);
-
-        $this->client->request(
-            method: 'POST',
-            uri: '/api/v2/admin/promotions',
-            server: $header,
-            content: json_encode([
-                'name' => 'T-Shirts discount',
-                'code' => 'tshirts_discount',
-                'startsAt' => '2023-12-04 12:30:00',
-                'endsAt' => '2023-11-04 12:30:00',
-            ], \JSON_THROW_ON_ERROR),
-        );
-
-        $this->assertResponse(
-            $this->client->getResponse(),
-            'admin/promotion/post_promotion_with_invalid_dates_response',
-            Response::HTTP_UNPROCESSABLE_ENTITY,
-        );
-    }
-
-    /** @test */
-    public function it_does_not_create_a_promotion_with_invalid_rules(): void
-    {
-        $this->loadFixturesFromFiles(['authentication/api_administrator.yaml', 'promotion/channel.yaml']);
-        $header = array_merge($this->logInAdminUser('api@example.com'), self::CONTENT_TYPE_HEADER);
-
-        $this->client->request(
-            method: 'POST',
-            uri: '/api/v2/admin/promotions',
-            server: $header,
-            content: json_encode([
-                'name' => 'T-Shirts discount',
-                'code' => 'tshirts_discount',
-                'rules' => [
-                    [
-                        'type' => CartQuantityRuleChecker::TYPE,
-                        'configuration' => [
-                            'count' => 'invalid',
-                        ],
-                    ],
-                    [
-                        'type' => CustomerGroupRuleChecker::TYPE,
-                        'configuration' => [
-                            'group_code' => 'wrong',
-                        ],
-                    ],
-                    [
-                        'type' => NthOrderRuleChecker::TYPE,
-                        'configuration' => [
-                            'nth' => 'invalid',
-                        ],
-                    ],
-                    [
-                        'type' => ShippingCountryRuleChecker::TYPE,
-                        'configuration' => [
-                            'country' => 'wrong',
-                        ],
-                    ],
-                    [
-                        'type' => HasTaxonRuleChecker::TYPE,
-                        'configuration' => [
-                            'taxons' => ['wrong'],
-                        ],
-                    ],
-                    [
-                        'type' => TotalOfItemsFromTaxonRuleChecker::TYPE,
-                        'configuration' => [
-                            'WEB' => [
-                                'taxon' => 'wrong',
-                                'amount' => 'invalid',
-                            ],
-                        ],
-                    ],
-                    [
-                        'type' => ContainsProductRuleChecker::TYPE,
-                        'configuration' => [
-                            'product_code' => 'wrong',
-                        ],
-                    ],
-                    [
-                        'type' => ItemTotalRuleChecker::TYPE,
-                        'configuration' => [
-                        ],
-                    ],
-                    [
-                        'type' => 'wrong_type',
-                        'configuration' => [
-                        ],
-                    ],
-                ],
-            ], \JSON_THROW_ON_ERROR),
-        );
-
-        $this->assertResponse(
-            $this->client->getResponse(),
-            'admin/promotion/post_promotion_with_invalid_rules_response',
-            Response::HTTP_UNPROCESSABLE_ENTITY,
-        );
-    }
-
-    /** @test */
-    public function it_does_not_create_a_promotion_with_invalid_actions(): void
-    {
-        $this->loadFixturesFromFiles(['authentication/api_administrator.yaml', 'promotion/channel.yaml']);
-        $header = array_merge($this->logInAdminUser('api@example.com'), self::CONTENT_TYPE_HEADER);
-
-        $this->client->request(
-            method: 'POST',
-            uri: '/api/v2/admin/promotions',
-            server: $header,
-            content: json_encode([
-                'name' => 'T-Shirts discount',
-                'code' => 'tshirts_discount',
-                'actions' => [
-                    [
-                        'type' => FixedDiscountPromotionActionCommand::TYPE,
-                        'configuration' => [
-                            'WEB' => [
-                                'amount' => 'invalid',
-                            ],
-                        ],
-                    ],
-                    [
-                        'type' => UnitFixedDiscountPromotionActionCommand::TYPE,
-                        'configuration' => [
-                            'WEB' => [
-                                'filters' => 'invalid',
-                            ],
-                        ],
-                    ],
-                    [
-                        'type' => PercentageDiscountPromotionActionCommand::TYPE,
-                        'configuration' => [
-                            'percentage' => -5,
-                        ],
-                    ],
-                    [
-                        'type' => UnitPercentageDiscountPromotionActionCommand::TYPE,
-                        'configuration' => [
-                            'WEB' => [
-                                'percentage' => 110,
-                            ],
-                        ],
-                    ],
-                    [
-                        'type' => ShippingPercentageDiscountPromotionActionCommand::TYPE,
-                        'configuration' => [
-                        ],
-                    ],
-                    [
-                        'type' => 'wrong_type',
-                        'configuration' => [
-                        ],
-                    ],
-                ],
-            ], \JSON_THROW_ON_ERROR),
-        );
-
-        $this->assertResponse(
-            $this->client->getResponse(),
-            'admin/promotion/post_promotion_with_invalid_actions_response',
-            Response::HTTP_UNPROCESSABLE_ENTITY,
-        );
-    }
-
-    /** @test */
     public function it_updates_promotion(): void
     {
         $fixtures = $this->loadFixturesFromFiles(['authentication/api_administrator.yaml', 'channel.yaml', 'promotion/promotion.yaml']);
